I've implemented a possible partial fix for the "update-landscape, update-memories problem" by letting the discovery (new child arguments, new in-/out attacks) happen in ticks mod 5 = 4 (so round 4, 9, 14,19…) which would mean agents cannot move away before obtaining this information b/c the order in round 4 would be:
 -   move-around
 -   update-landscape
 -   compute-popularity

and then comes round 5:

 -   update memories
    […]

so agents should have a updated memory of those big changes in landscape which happen every 5th round. But this still doesn't fix the problem when arguments turn red as this might happen on any round and if not the next round is an update-memory round (mod 5 = 0) then the agent will move away before learning the information revealed by the argument turning red.