Clean up of interface and order change in go
1. Added an agents-own variable, needed for the clean up of the behavior
file
2. Removed the switch col-networks, since we don't use it for the KRR
paper
3. Just one theory gives problems: the minimum is now 2

4. The biggest change: I've changed the order of the procedures in the
go procedure.
a) The idea is now that agents start every time step with looking around
and updating their memory with their current neighborhood.
b) Every five time steps agents share information with the agents from
their own network, this can of course be changed into every round, but
that seems a bit much.
c) With the probability from the interface, agents share with an agent
from another network, undirected or directed and only the information
they want to share.
d) Like before agents compute the current best theory and act on that
information every five time steps.
e) Every time step (either after sharing and the strategy if it is a 5th
time step and otherwise after the memory update of the current
neighborhood) the agent moves around and updates the landscape.