Started with clean up of behavior.nls
For this the variable neighborargs had to be changed a bit, it is now an
agents-own variable. In order to update the personal memory of the
agents every time after the moving around and update of the landscape
the update-memories procedure is split:

1. update-memories: it updates the personal memories with the current
neighborhood information.
2. share-with-others: procedure with subprocedures that is only executed
once every five rounds; First agents share in their own network their
whole personal memory. Then two subprocedures follow:
a) create-share-memory: in which the agents collect the info they are
prepared to share with agents that do not belong to their own network;
b) share-with-other-network: in which the actual sharing takes place.

In the move-around procedure a subprocedure is created, that runs
immediately for all agents that work on a not fully researched
not-admissible argument: find-defense, it is the procedure we still have
to look into with the idea that it lets agents look for a defense for
their current attacked argument.

The changes of the landscape that occur because an argument is fully
discovered is now a separate procedure: full-discovery.
The  use of the color turquoise is changed according to Daniels
comments.