Support -> Discovery, update of comments
In the whole program support relations are changed into discovery
relations, in order to let the program correspond with the KRR paper.
One or two instances might be overlooked.

Comments for local variables are updated (for the behavior.nls and
strategies.nls files), as well as some of the general comments. All
comments should be check before publishing.

In behavior.nls some variables were not clear (e.g. add-argu was about
relations instead of arguments), this should be better now.