Prospective movement fix
Based on Daniel's code the find-defense procedure is updated. The idea
is still the same: agents working on a not-admissible not fully
researched argument try to find among their disovered child arguments a
(not yet discovered) defense: an attack from that discovered child to
the attacking argument.

Agents can only move once, this is now a variable of the agents. This is
necessary because of the find-defense procedure, in which agents might
move to an admissible argument and they might later, in the same time
step, be able (because of the probabilities) to move to another
argument.

In the update-memories procedure a link-entry was still "s" for
discovery relations, instead of "d", this is now fixed.

Arguments (starts and arguments) have an extra variable, in which they
keep track of the moment that they are fully researched. If at a certain
time point the argument turns red, this variable is set to true. In the
full-discovery procedure only arguments with this variable as true are
considered, hence each argument is considered at most once.