Initialization of new variables added
For starts and arguments the full-research variable is set to false. It
will be set to true in the time step in which the start/argument turns
red.

For agents the moved variable is set to false. Everytime an agent moves
(normal or prospective) it will be set to true. At the beginning of the
move-around procedure the variable is set to false for all agents.