Documentation update
The "full-research" variable was added for starts and arguments. The
"moved" variable was added for agents. In both cases the comments were
updated as well.

In the interface it is now possible to choose between "reliable" and
"biased" agents in the "social-actions" chooser, instead of "kind" and
"biased".

The documentation was updated with the resent decisions. The content of
the documentation was added to the readme file.