Jumping to an argument in new theory
Agents that are already aware of arguments in the theory they jump to,
jump to a random argument that they know about. Now this random argument
cannot be a turquoise argument.

Note that the argument might have changed color (is no longer turquoise)
in the objective landscape, but if the agent only has the argument as
turquoise in its memory it will not jump to it.