Discovery of new child-arguments
The discovery of new child-arguments depends now on the color (value of
yellow or lower) but per color only one child-argument can be
discovered. Meaning that, if an argument did not change color in the
last 5 time steps, it will not be the case that a new child-argument is
found.