First introduction of Zollman's social networks
This commit contains the first setup of introducing social structures
for the communication among agents from different collaborator-networks.
A short and incomplete summary:

* in the interface the user can choose between three social structures
for connecting the collaborator-networks: cycle; wheel and complete;

* there are three new agents-own variables: rep-agent (whether the agent
is the representative agent); and to-add-mem-argu and to-add-mem-rel
(lists with arguments resp. relations that the agent has to add to its
memory);

* there are two new global variables: colla-networks (a list with all
collaborator-networks in it); and share-structure (a list where each
entry is a list of collaborator-networks representing the networks the
first collaborator-network of that entry shares information with);

* after creating the collaborator networks and the memory of the agents
the social structures are created, these are saved in the
share-structure variable;

* only the representative agents create lists with arguments and
relations they want to share with the representative agents of
neighboring networks;

* the direction of sharing does not depend on a probability: it is
always bidirectional;

* based on the number of new arguments and relations the communication
costs are calculated. At most 30 new arguments/relations can be added to
the memory and at most 10 per day. Communication = 0 if number of new
entries = 0; communication = 1 if 0 < number of new entries < 11;
communication = 2 if 10 < number of new entries < 21; and communication
= 3 if 20 < number of new entries < 31;

* instead of true/false the communication variable is now a number (0,
1, 2 or 3). Every time step (except for ticks mod 5 = 0) the non-zero
communication values are decreased. On ticks mod 5 = 0 also the
communicating agents do research;

* in the run-many file the global variables are reset and the run-many
setup procedures are updated;

* the documentation and readme-file were updated as well.