Removal of chooser "sharing"
Because of the way the memory of the representative agents is updated,
it is not longer possible to let them share their whole network. It
would of course be possible to implement this, but that would increase
the time costs, also, we have had nice results when agents share only
their neighborhood and I think we agreed that this was more realistic
anyway.