Moment of cumminicating/ticks check
* At the beginning of the move-around and find-defense procedure it is
now checked whether the current agent is communicating and/or if it is a
zero round. Agents that have a non-zero value of communicating are not
able to run this code unless it is a zero round.

* Agents on red argument cannot move to arguments on which an agent of
their own collaborator network is already working. This was already the
case for some conditions, but this should now be more strict.