Preparations for update of the actual procedures
The old procedures are savet into inputfile-old so that they can still be run for comparison when the actual procedures are updated.
Those will only be part of the model in the testprocedures branch but not in the soon to be opened regular branch which will not contain all the stuff from the testprocedures.