Small performance upgrade to admissibility calculation
A new agentset within the compute-subjective attacked procedure (admissibility calculation) is created, which contains all subjective-arguments. This is done in order to make the sets which have to be checked repeatedly smaller and results in quite a performance gain (procedure approx twice as fast) with no changes to the structure.
Aditionally there has been a small change regarding the order of the conditions for some "if" and "with" statements within compute-subjective-attacked, to allow for faster computation.