Distributing the researchers in groups of five during the setup
The researchers are randomly distributed among the starts in groups of five which corresponds to the number of members each collaborative network has. This has been introduced to prevent singleton groups (with all their special needs) from occurring. This also removes an asymmetry between homogenous and heterogenous groups as the latter always had group sizes of five while in the former case this could vary from one to five.