Introduction of new global- and agents-own-variables
New variables have been introduced and documented. The documentation is still work in progress and somehow the format looks strange (I tried to make an unordered list with subitems, but in the Netlogo editor this looks totally strange). For details on the variables see the documentation in the model (last section: ## Variables). This variables are prerequisites for the procedures which will soon get transferred from the testbranch.