Transfer of behavior part1
* `move-to-nextargu` has been transferred.  This procedure is called by researchers when they move (to nextargu). It makes sure that the researcher has an updated memory of her surrounding before moving and updates the argument she's working on (mygps - cf. infotab) to the argument she's moving to.
* `update-memories` has been transferred and the corresponding infotab wip section has been updated to include the new information: Researchers no longer keep track of  discovery relations as separate entities as this information is implicit  in the knowledge of the argument itself.
The memory management is comprised of two parts:
(a) The researchers save arguments and relations in the form of turtle-sets / link-sets in their memory (cf. infotab Variables -> `to-add-mem-argu` `to-add-mem-rel`) which will be synchronized every week with the group in the `share-with-group` procedure
(b) the status in which the argument / relation is known to a certain collaborative network (=group) is saved in the argument / link itself.  (cf. infotab Variables -> `group-color-mem`, `in-group-i-memory`). For links this will be facilitated during the `share-with-group` procedure, while for arguments the color is updated right when the researchers update their memory

* New information on `to-add-mem-argu` `to-add-mem-rel` in the infotab wip section has been added as those variables are utilized differently from before
* Additionally some cosmetic changes to the wip section of the infotab, hopefully subitems in the lists are shown in a somewhat more readeable manner now