Fix for prospective movement
* The new update-memories reduced the update frequency to: every 5th round or when researchers were moving.
The problem with this in the `find-defense` procedure is that this might lead the researchers, who are looking for a defense, to ignore an argument which very recently popped up in the landscape.
In order to fix this the group of the current researcher now updates their memory during the find-defense procedure, to see whether they can find a defender in their neighborhood.