Change from group to agent communal network
The way in which agents from different collaborative networks communicate with each other is changed. Instead of sharing with fixed agents from other networks each agent shares with a certain probability information with a random agent from another network. Additionally this communication can be directional (only the randomly chosen agent receives information) or bidirectional in which information is exchanged.